+++ scripts/after-install.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+
+export PATH=/bin:/sbin:/usr/bin:/usr/sbin
+echo -n "Adding metcap user and group... "
+useradd -r -d /etc/metcap -s /bin/false -U metcap
+echo "done."